Add more lecture 5 readings
This changeset adds two more reading assignments to lecture 5.

* Vaughn's post is an excellent high-level view of how a blockchain for
patient data could work. What makes it notable is that it avoids several
pitfalls and antipatterns we usually see in blockchain-related
proposals.

* Kessler's article reinforces the points that Kish & Topol make. It
also points to another stakeholder that could make this health data
economy a reality: insurance companies. (The article sits behind a
paywall but should be accessible to the students via the school's
library.)

Signed-off-by: Kostas Christidis <kostas@christidis.io>